resolved linter errors
