+++ apps/web/src/components/GoToCommunity/BuildersMostWanted.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import classNames from 'classnames';
 import {
   gridClasses,
