+++ apps/web/src/components/GoToCommunity/Funding.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   sectionContainer,
   titleClasses,
