+++ apps/web/src/components/GoToCommunity/GetNoticed.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import {
   sectionContainer,
   titleClasses,
