+++ apps/web/src/components/GoToCommunity/StartBuilding.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import classNames from 'classnames';
 import {
   sectionContainer,
